requirements.txt
python-dotenv
pandas
numpy
pymango
certifi
setuptools
requests
pymongo[srv]
scikit-learn
dill
pyaml
mlflow

## -e .